Update FPSController.gd
Basically a small fix to ensure the wall we're checking isn't completely vertical (or basically vertical).  I was finding that by jumping into walls, the player would hover above the ground and never technically be on the ground because they're in floating mode.  

This fixes that allowing you to jump into walls as you please.